+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [not released]
 
+### Added
+ - Added support for `breakpoint` statement
+
 ### Changed
  - Fixed `max-public-methods` linter check disabling (#222)
  - Default regex for names of constants now allows underscore as a prefix to denote private contants (#223)